+++ 1.21.4/src/main/java/io/github/axolotlclient/modules/hud/gui/hud/item/ArmorHud.java
@@ -117,7 +117,7 @@ public class ArmorHud extends TextHudEntry implements DynamicallyPositionable {
 	private void renderDurabilityNumber(GuiGraphics graphics, ItemStack stack, int x, int y) {
 		boolean showDurability = showDurabilityNumber.get();
 		boolean showMaxDurability = showMaxDurabilityNumber.get();
-		if (!(showMaxDurability || showDurability) || stack.isEmpty()) {
+		if (!(showMaxDurability || showDurability) || stack.isEmpty() || stack.getMaxDamage() == 0) {
 			return;
 		}
 		String text = showDurability && showMaxDurability ? (stack.getMaxDamage() - stack.getDamageValue())+"/"+stack.getMaxDamage() : String.valueOf((showDurability ? stack.getMaxDamage() - stack.getDamageValue() : stack.getMaxDamage()));;