+++ 1.8.9/src/main/java/io/github/axolotlclient/mixin/ChatHudAccessor.java
@@ -40,4 +40,7 @@ public interface ChatHudAccessor {
 
 	@Accessor("trimmedMessages")
 	List<ChatMessage> getVisibleMessages();
+
+	@Accessor("hasNewMessagesSinceScroll")
+	boolean hasUnreadNewMessages();
 }